全项目测试前的第一次提交

+++ include/Motor.h
@@ -12,13 +12,13 @@ private:
     uint8_t pin_m1_B;
     uint8_t pin_m2_A;
     uint8_t pin_m2_B;
-    uint8_t motor1Speed;
-    uint8_t motor2Speed;
+    int32_t motor1Speed;
+    int32_t motor2Speed;
 
 public:
     Motor(uint8_t pin_m1_A, uint8_t pin_m1_B, uint8_t pin_m2_A, uint8_t pin_m2_B);
     void setup();
-    void setSpeed(uint8_t motor, uint8_t speed);
+    void setSpeed(uint8_t motor, int32_t speed);
     void setPins(uint8_t pin_m1_A, uint8_t pin_m1_B, uint8_t pin_m2_A, uint8_t pin_m2_B);
     void update() const;
 };
